Factor out suite client creation in tests
