+++ tests/Integration/Suite/AbstractSuiteTestCase.php
@@ -1,0 +1,33 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SmartAssert\ApiClient\Tests\Integration\Suite;
+
+use GuzzleHttp\Client as HttpClient;
+use GuzzleHttp\Psr7\HttpFactory;
+use SmartAssert\ApiClient\Exception\Error\Factory as ExceptionFactory;
+use SmartAssert\ApiClient\Factory\Source\SuiteFactory;
+use SmartAssert\ApiClient\ServiceClient\HttpHandler;
+use SmartAssert\ApiClient\SuiteClient;
+use SmartAssert\ApiClient\Tests\Integration\AbstractIntegrationTestCase;
+
+abstract class AbstractSuiteTestCase extends AbstractIntegrationTestCase
+{
+    protected static SuiteClient $suiteClient;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        self::$suiteClient = new SuiteClient(
+            new SuiteFactory(),
+            new HttpHandler(
+                new HttpClient(),
+                new ExceptionFactory(self::$errorDeserializer),
+                new HttpFactory(),
+                self::$urlGenerator,
+            ),
+        );
+    }
+}
